Revert "move react-switch"
This reverts commit a4367ed5f48e04520dd5a46be88bf7b9113e6ad7.
