+++ packages/react-switch/.babelrc.json
