+++ packages/react-switch/.eslintrc.json
