+++ packages/react-switch/.npmignore
