+++ packages/react-switch/.storybook/main.js
@@ -1,6 +1,6 @@
-const rootMain = require('../../../../.storybook/main');
+const rootMain = require('../../../.storybook/main');
 
-module.exports = /** @type {Omit<import('../../../../.storybook/main'), 'typescript'|'babel'>} */ ({
+module.exports = /** @type {Omit<import('../../../.storybook/main'), 'typescript'|'babel'>} */ ({
   ...rootMain,
   stories: [...rootMain.stories, '../src/**/*.stories.mdx', '../src/**/*.stories.@(ts|tsx)'],
   addons: [...rootMain.addons],
