+++ packages/react-switch/LICENSE
