+++ packages/react-switch/MIGRATION.md
