+++ packages/react-switch/Spec.md
