+++ packages/react-switch/bundle-size/Switch.fixture.js
