+++ packages/react-switch/config/api-extractor.json
