+++ packages/react-switch/config/tests.js
