+++ packages/react-switch/jest.config.js
@@ -5,7 +5,7 @@
  */
 module.exports = {
   displayName: 'react-switch',
-  preset: '../../../jest.preset.js',
+  preset: '../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsConfig: '<rootDir>/tsconfig.spec.json',
