+++ packages/react-switch/src/Switch.ts
