+++ packages/react-switch/src/common/isConformant.ts
