+++ packages/react-switch/src/components/Switch/Switch.test.tsx
