+++ packages/react-switch/src/components/Switch/Switch.types.ts
