+++ packages/react-switch/src/components/Switch/index.ts
