+++ packages/react-switch/src/components/Switch/renderSwitch.tsx
