+++ packages/react-switch/src/components/Switch/useSwitchStyles.ts
