+++ packages/react-switch/src/stories/SwitchBestPractices.md
