+++ packages/react-switch/src/stories/SwitchChecked.stories.tsx
