+++ packages/react-switch/src/stories/SwitchLabel.stories.tsx
