+++ packages/react-switch/src/stories/SwitchLabelWrapping.stories.tsx
