+++ packages/react-switch/src/stories/SwitchRequired.stories.tsx
