+++ packages/react-switch/tsconfig.lib.json
