+++ packages/react-switch/tsconfig.spec.json
