+++ tsconfig.base.json
@@ -53,7 +53,7 @@
       "@fluentui/react-spinner": ["packages/react-components/react-spinner/src/index.ts"],
       "@fluentui/react-storybook": ["packages/react-components/react-storybook/src/index.ts"],
       "@fluentui/react-storybook-addon": ["packages/react-components/react-storybook-addon/src/index.ts"],
-      "@fluentui/react-switch": ["packages/react-components/react-switch/src/index.ts"],
+      "@fluentui/react-switch": ["packages/react-switch/src/index.ts"],
       "@fluentui/react-tabs": ["packages/react-components/react-tabs/src/index.ts"],
       "@fluentui/react-tabster": ["packages/react-components/react-tabster/src/index.ts"],
       "@fluentui/react-text": ["packages/react-components/react-text/src/index.ts"],
